refactor: split exerciseEntryHistoryImport into custom hook
